requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
requests==2.31.0
python-dotenv==1.0.0